+++ src/FolkerKinzel.Strings.Tests/StringBuilderExtensionTests.cs
@@ -283,6 +283,7 @@ public class StringBuilderExtensionTests
     [TestMethod]
     public void IndexOfTest1() => Assert.AreEqual(1, new StringBuilder("testen").IndexOf('e'));
 
+
     [TestMethod()]
     [ExpectedException(typeof(ArgumentNullException))]
     public void IndexOfTest2()
@@ -1324,12 +1325,12 @@ public class StringBuilderExtensionTests
     public void AppendBase64Test2()
     {
         StringBuilder? builder = null;
-        _ = builder!.AppendBase64(new byte[] { 42 }.ToList());
+        _ = builder!.AppendBase64("*"u8.ToArray().ToList());
     }
 
-    [TestMethod]
-    [ExpectedException(typeof(ArgumentNullException))]
-    public void AppendBase64Test3() => _ = new StringBuilder().AppendBase64((List<byte>)null!);
+    //[TestMethod]
+    //[ExpectedException(typeof(ArgumentNullException))]
+    //public void AppendBase64Test3() => _ = new StringBuilder().AppendBase64((List<byte>)null!);
 
 
     [TestMethod]
@@ -1342,21 +1343,21 @@ public class StringBuilderExtensionTests
     [TestMethod]
     public void AppendBase64Test5()
     {
-        StringBuilder builder = new StringBuilder().AppendBase64(new byte[] { 42 }, Base64FormattingOptions.InsertLineBreaks);
+        StringBuilder builder = new StringBuilder().AppendBase64("*"u8.ToArray(), Base64FormattingOptions.InsertLineBreaks);
         Assert.IsFalse(builder.Contains('\n'));
     }
 
     [TestMethod]
     public void AppendBase64Test6()
     {
-        StringBuilder builder = new StringBuilder(new string('a', 75)).AppendBase64(new byte[] { 42 }.AsSpan(), Base64FormattingOptions.InsertLineBreaks);
+        StringBuilder builder = new StringBuilder(new string('a', 75)).AppendBase64("*"u8.ToArray().AsSpan(), Base64FormattingOptions.InsertLineBreaks);
         Assert.IsTrue(builder.Contains('\n'));
     }
 
     [TestMethod]
     public void AppendBase64Test7()
     {
-        StringBuilder builder = new StringBuilder(new string('a', 77)).AppendBase64(new byte[] { 42 }.AsSpan(), Base64FormattingOptions.InsertLineBreaks);
+        StringBuilder builder = new StringBuilder(new string('a', 77)).AppendBase64("*"u8.ToArray().AsSpan(), Base64FormattingOptions.InsertLineBreaks);
         Assert.IsTrue(builder.Contains('\n'));
     }
 
@@ -1373,6 +1374,5 @@ public class StringBuilderExtensionTests
 
 
     [TestMethod]
-    [ExpectedException(typeof(ArgumentNullException))]
-    public void AppendBase64Test9() => new StringBuilder().AppendBase64((byte[]?)null!);
+    public void AppendBase64Test9() => Assert.AreEqual(0, new StringBuilder().AppendBase64((IEnumerable<byte>?)null).Length);
 }