+++ src/FolkerKinzel.Strings/StringBuilderExtension_AppendBase64.cs
@@ -10,7 +10,7 @@ public static partial class StringBuilderExtension
     /// am Ende eines <see cref="StringBuilder"/>-Objekts an.
     /// </summary>
     /// <param name="builder">Der <see cref="StringBuilder"/>, an den Zeichen angefügt werden.</param>
-    /// <param name="bytes">Die <see cref="byte"/>-Enumeration, die die anzufügenden Daten enthält.</param>
+    /// <param name="bytes">Die <see cref="byte"/>-Enumeration, die die anzufügenden Daten enthält oder <c>null</c>.</param>
     /// <param name="options">Ein Enumerationswert, der es erlaubt festzulegen, ob in das Base64 automatisch Zeilenumbrüche 
     /// eingefügt werden sollen.</param>
     /// <returns>Ein Verweis auf <paramref name="builder"/>, nachdem der Anfügevorgang abgeschlossen wurde.</returns>
@@ -19,12 +19,12 @@ public static partial class StringBuilderExtension
     /// <exception cref="ArgumentNullException"><paramref name="builder"/> oder <paramref name="bytes"/> ist <c>null</c>.</exception>
     /// <exception cref="ArgumentOutOfRangeException">Bei der Erhöhung der Kapazität von <paramref name="builder"/>
     /// würde <see cref="StringBuilder.MaxCapacity"/> überschritten.</exception>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static StringBuilder AppendBase64(this StringBuilder builder,
-                                             IEnumerable<byte> bytes,
-                                             Base64FormattingOptions options = Base64FormattingOptions.None) =>
-        bytes is null
-               ? throw new ArgumentNullException(nameof(bytes))
-               : builder.AppendBase64(CollectionConverter.AsReadOnlySpan(bytes), options);
+                                             IEnumerable<byte>? bytes,
+                                             Base64FormattingOptions options = Base64FormattingOptions.None)
+        => bytes is null ? builder
+                         : Base64.AppendEncodedTo(builder, CollectionConverter.AsReadOnlySpan(bytes), options);
 
 
     /// <summary>
@@ -32,19 +32,20 @@ public static partial class StringBuilderExtension
     /// am Ende eines <see cref="StringBuilder"/>-Objekts an.
     /// </summary>
     /// <param name="builder">Der <see cref="StringBuilder"/>, an den Zeichen angefügt werden.</param>
-    /// <param name="bytes">Das <see cref="byte"/>-Array, das die anzufügenden Daten enthält.</param>
+    /// <param name="bytes">Das <see cref="byte"/>-Array, das die anzufügenden Daten enthält oder <c>null</c>.</param>
     /// <param name="options">Ein Enumerationswert, der es erlaubt festzulegen, ob in das Base64 automatisch Zeilenumbrüche 
     /// eingefügt werden sollen.</param>
     /// <returns>Ein Verweis auf <paramref name="builder"/>, nachdem der Anfügevorgang abgeschlossen wurde.</returns>
     /// <remarks>Die Methode verwendet eine eigene Base64-Implementierung, die etwas langsamer als die BCL-Methoden
     /// ist, aber für den Zweck sehr viel weniger Heap-Speicher alloziert.</remarks>
-    /// <exception cref="ArgumentNullException"><paramref name="builder"/> oder <paramref name="bytes"/> ist <c>null</c>.</exception>
+    /// <exception cref="ArgumentNullException"><paramref name="builder"/> ist <c>null</c>.</exception>
     /// <exception cref="ArgumentOutOfRangeException">Bei der Erhöhung der Kapazität von <paramref name="builder"/>
     /// würde <see cref="StringBuilder.MaxCapacity"/> überschritten.</exception>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static StringBuilder AppendBase64(this StringBuilder builder,
-                                             byte[] bytes,
-                                             Base64FormattingOptions options = Base64FormattingOptions.None) =>
-        builder.AppendBase64(bytes is null ? throw new ArgumentNullException(nameof(bytes)) : bytes.AsSpan(), options);
+                                             byte[]? bytes,
+                                             Base64FormattingOptions options = Base64FormattingOptions.None) 
+        => Base64.AppendEncodedTo(builder, bytes.AsSpan(), options);
 
 
     /// <summary>
@@ -61,10 +62,10 @@ public static partial class StringBuilderExtension
     /// <exception cref="ArgumentNullException"><paramref name="builder"/> ist <c>null</c>.</exception>
     /// <exception cref="ArgumentOutOfRangeException">Bei der Erhöhung der Kapazität von <paramref name="builder"/>
     /// würde <see cref="StringBuilder.MaxCapacity"/> überschritten.</exception>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static StringBuilder AppendBase64(this StringBuilder builder,
                                              ReadOnlySpan<byte> bytes,
-                                             Base64FormattingOptions options = Base64FormattingOptions.None) =>
-         builder is null ? throw new ArgumentNullException(nameof(builder)) 
-                         : Base64.AppendEncodedTo(builder, bytes, options);
+                                             Base64FormattingOptions options = Base64FormattingOptions.None)
+        => Base64.AppendEncodedTo(builder, bytes, options);
     
 }